change it for a stable hashing algo
