+++ engine/schema/schemeable.test.ts
@@ -77,7 +77,7 @@ Deno.test("Simple type generation", async () => {
     fail("SimpleType should exists");
   }
 
-  const name = Deno.build.os === "windows" ? "tl@157-178" : "tl@155-174";
+  const name = "tl@155-174";
   assertEquals(transformed, {
     file: path,
     type: "alias",
@@ -153,7 +153,7 @@ Deno.test("TwoRefsProperties type generation", async () => {
     fail("TwoRefsProperties should exists");
   }
 
-  const name = Deno.build.os === "windows" ? "tl@720-790" : "tl@682-749";
+  const name = "tl@682-749";
   assertObjectMatch(transformed, {
     "type": "alias",
     "jsDocSchema": {},
